More custom rooms...
* Added more custom rooms courtasy of Neighborino. (all new rooms for
Gungeon Proper).

* Verious other adjustments like tweaks to custom boss's HP on my custom
secret floor.

* Tweaks to sound fx for destroying glithed enemies produced by using
the Corruption Bomb item. It should hopefully have the correct volume
now.

* Changed room category of "Bootleg" room from "normal" to "connector".
This should increase odds of it appearing in a run.

* Added another custom trap room design.

* Fixed bug with Alarm Mushrooms that could cause soft locks. My code
for placing them was erroneously placing them in trap rooms that had no
enemies. If a player triggered one in this room, it would not unseal
after the player clears the room. I have refactored this to instead
check that a room contains combat enemies before placing alarm
mushrooms. Should resolve this issue.