1.0 Release Commit...
* 100+ new rooms from TheTurtleMelon has been added.
* Special trap room no longer spawns enemies if it shows up as a secret
room due room clear events not working in secret rooms. (the room would
never unseal)
* The Lead Key can now show up as an item in the Key shop and the Baby
Good Hammer can show up as an item in the Blacksmith on the Forge.
* Odds for Raccoon and Turkey junk enemies spawning have been reduced.
Odds of getting double junk from Raccoon also reduced.
* Glitch Elevator can no longer appear in the Bootleg room where it
would be inaccessible to the player.
* Another new custom item + custom item synergy added. "The Lead Key". A
special active item that will take the player to a random new room not
connected to the rest of the floor. The custom synergys added is Master
of Unlocking. A synergy betweeen Trusty Lock Picks and The Lead Key.
Having this synergy gives the lock picks a 100% success rate. Having the
Master of Unlocking passive also enables this synergy. Thanks to
Nevernamed for the sprite and item idea!
* Bootleg guns have new feature. Chance to transmorgify enemies into
bootleg versions. Guns now have a chance of exploding while taking
damage. Bootleg guns will explode if they run out of ammo due to
questionable build quality. Exploded guns are permanently removed from
player's inventory. The guns exploding will damage the player. This
makes them deserving of the D teir they are in.
* Expand_HopScotch has been disabled until re-balancing due to
difficulty in clearing this trap room.