Change how WestBros boss handles animations...
Changed it to interact directly with the sprite animator instead of
using AIAnimator for the boss intro. AIAnimator will transition the
AIActor back to it's idle animation when it is finished playing any
specified clip given to it.

This caused issues trying to correctly time the code that enables the
gun/gun hands after the intro animation.

Also added new settings to ExpandTheGungeon_Settings that allows setting
a test floor/dungeon flow on startup for quick testing of stuff. :D