New expirimental feature for disabling GC...
New console command added that can allow disabling GC and only running
it when floor reloads or if a certain maximum memory allowance is
reached. Note that is recommended you have more then 8GB of ram for this
feature to work best. It is off by default. use the command "expand
togglegc" to enable this setting. This setting is saved to settings file
and will stay on for future game restarts until you use the command
again to turn it off.
* Fix incorrect starting sprite for Gungeoneer Mimic's mirror from
previous commit after the conversion to serialized sprite collections.