v2.9.5 Release Commit...
* Removed switch object from floor 1 entrance room as it was causing
confusion. The default state for the music switch toggle for Rickroll
chests is now "off" and can only be changed in the settings file now.
Players who have already enabled it will not need to re-enable it.
* TheLeadKey can no longer teleport the player to an inaccessible area
in Winchester rooms and should be a playable winchester room.
* TheLeadKey will automatically remove enemies that overlap the area the
player will be teleported to which should cut down on players taking
unavoidable contact damage.
* One of the custom secret room layouts for the Witch NPC resulted in
guns spawning in a hard to reach area or glitching into the wall. The
layout has been adjusted to address this issue.