2.11.5 Release Commit...
* Disable custom glitch screen shaders for The Lead Key, glitch chest
floors, etc if end user forces game to use OpenGL or Vulkan backends as
this appears to break the shaders.
* Custom Shaders moved to seperate asset bundle to make mod more file
space efficient. MacOS version of shader bundle now available along with
the Linux version. Custom shaders should now in theory work on MacOS for
the ... 3 people that got mods working there. :P
* Shaders moved to sepeate Unity Editor project to avoid incompatiblity
with scripts that that prevented asset bundles from being built for
MacOS. (part of the fix that allows working shader bundle for MacOS)
* AssetBundle folders in Unity Editor projects now added to Github
gitignore.
* More improvements to code that places Alarm Mushrooms/Jungle trees.