Prepare for 3.0 Update...
* New item: Mr Cap. Sprite work courtasy of Dallan. New active item that
is a surperior version of Third Party Controller.
* Asset loading overhaul. Most of the mod is now loaded async via
IEnumerator. This is due to the introduction of custom loading screens!
Even the one seen when the game starts up is now customized. It can now
detail current load progress of the mod! Customized load screens for
certain custom floors are also planned!